% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.Cochran.Armitage.trend.exact.2x3.R
\name{my.Cochran.Armitage.trend.exact.2x3}
\alias{my.Cochran.Armitage.trend.exact.2x3}
\title{All possible 2x3 tables
This function returns nx6 matrix; each row has 6 elements of 2x3 table
sharing marginal counts with argument m.}
\usage{
my.Cochran.Armitage.trend.exact.2x3(m)
}
\arguments{
\item{m}{is a vector with 6 non-negative values, the first 3 values for cases}
}
\description{
All possible 2x3 tables
This function returns nx6 matrix; each row has 6 elements of 2x3 table
sharing marginal counts with argument m.
}
\examples{
m <- c(9,42,49,6,43,51)
my.Cochran.Armitage.trend.exact.2x3(m)
my.trend.exact.2x3(m)
}
\keyword{SNPs}
\keyword{exact}
\keyword{trend}

