% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.doubles.bad.R
\name{my.doubles.bad}
\alias{my.doubles.bad}
\title{Score Badminton Doubles Game}
\usage{
my.doubles.bad(K, p = 21)
}
\arguments{
\item{K}{is a vector of 1 or 2.}

\item{p}{is the max point of a game.}
}
\description{
This function returns players' info from a serial record of pointing teams.
}
\examples{
K <- sample(1:2,42,replace=TRUE,prob=c(0.49,0.51))
K <- c(1,1,1,1,1,1,2,1,2,1,2,1,1,2,1,2,2,1,1,2,2,1,1,1,2,1,2,1,1,2,1,1)
K <- K * (-1) + 3
# The score sheet for this match : http://www.badminton.or.jp/nba/shinpan/scoresheet.pdf
doubles.score <- my.doubles.bad(K)
# ポイントチーム記録
doubles.score$Point.team
# チーム別累積ポイントログ
doubles.score$Points
# サーバーは誰か
doubles.score$Server.4
# サーバー・レシーバーの左右位置
doubles.score$Server.side
# 各チームのメンバーの位置(初期位置かその逆か)
doubles.score$Location
}
\keyword{miscelaneous}

