% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.EulTriSph.R
\name{my.EulTriSph}
\alias{my.EulTriSph}
\title{Dummy title}
\usage{
my.EulTriSph(n.step)
}
\description{
Dummy description.
}
\examples{
n.step <- 20
euler.out <- my.EulTriSph(n.step)
el <- my.dir.graph(euler.out[[1]],euler.out[[2]])
plot(graph.edgelist(el))
root.e <- el[1,]
euler.tree.out <- my.euler.tree(euler.out[[1]],euler.out[[2]],root.e)
g.tree <- graph.edgelist(euler.tree.out[[1]])
is.connected(g.tree)
g.tree
sum(euler.out[[2]])
nrow(euler.out[[1]])
sh.dist <- shortest.paths(graph.edgelist(euler.tree.out[[2]]),root.e[1],mode="out")
plot(graph.edgelist(euler.tree.out[[1]]),vertex.label=sh.dist)
}

