% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.grid.adjacent.R
\name{my.grid.adjacent}
\alias{my.grid.adjacent}
\title{2D grid graph adjacency matrix
This function returns an adjacent matrix of 2D grid graph}
\usage{
my.grid.adjacent(n, m = n)
}
\arguments{
\item{n}{(m) is an integer indicating grid size}
}
\description{
2D grid graph adjacency matrix
This function returns an adjacent matrix of 2D grid graph
}
\examples{
library(igraph)
out <- my.grid.adjacent(3,4)
image(out)
plot(graph.adjacency(out))
}
\keyword{graph}
\keyword{grid}

