% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_complex_color.R
\name{my_complex_color}
\alias{my_complex_color}
\title{Schwarz-Christoffel-like curve}
\usage{
my_complex_color(z, int0 = 0.6, sat0 = 0.3, int1 = 1, sat1 = 1)
}
\arguments{
\item{p}{is a positive real number.}

\item{n}{is a positive integer; No. steps.}

\item{S0}{is a complex number with 0 imaginary.}

\item{p2}{is a positive real number >= p.}
}
\description{
This function gives Schwarz-Christoffel-like curve.
}
\examples{
out <- my_complex_color(3+1i*2)
}
\keyword{Schwarz-Crhistoffel}

