% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/my_schwarzChristoffel.R
\name{my_schwarzChristoffel}
\alias{my_schwarzChristoffel}
\title{Schwarz-Christoffel-like curve}
\usage{
my_schwarzChristoffel(p, n, S0 = 0 + 0 * (0+1i), p2 = NULL)
}
\arguments{
\item{p}{is a positive real number.}

\item{n}{is a positive integer; No. steps.}

\item{S0}{is a complex number with 0 imaginary.}

\item{p2}{is a positive real number >= p.}
}
\description{
This function gives Schwarz-Christoffel-like curve.
}
\examples{
out <- my_schwarzChristoffel(0.2,30)
plot(out)
}
\keyword{Schwarz-Crhistoffel}

