% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_sphere_tri_mesh.R
\name{my_sphere_tri_mesh}
\alias{my_sphere_tri_mesh}
\title{Sphere Triangular Mesh}
\usage{
my_sphere_tri_mesh(n.psi = 30)
}
\arguments{
\item{n.psi}{is an integer.}
}
\description{
This function gives 3D coordinates of triangular mesh on sphere..
}
\examples{
sp.mesh <- my_sphere_tri_mesh(n.psi=30)
library(rgl)
plot3d(sp.mesh$xyz)
segments3d(sp.mesh$xyz[c(t(sp.mesh$edge)),])
}
\keyword{mesh}
\keyword{sphere}
\keyword{triangulation}

