% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ry_ARGsim.R
\name{ry_ARGsim}
\alias{ry_ARGsim}
\title{Ancestral Recombination Graph simulation}
\usage{
ry_ARGsim(n = 8, g = 5, l = 6, r = r)
}
\arguments{
\item{n}{is a No. of initial members}

\item{g}{is a No. generations.}

\item{l}{is a parameter}

\item{r}{is another paramter.}
}
\description{
This function makes a data set under the model of ARG
}
\examples{
n<-8
g<-10
l<-200
r<-0.05
d<-ry_ARGsim(n=n,g=g,l=l,r=r)
mmm<-matrix(d$parents[1,],ncol=g)
}
\keyword{population_genetics}

