% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ry_hweExact.R
\name{ry_hweExact}
\alias{ry_hweExact}
\title{Hardy-Weinberg Exact Test}
\usage{
ry_hweExact(g = c(813, 182, 5))
}
\arguments{
\item{g}{is a integer vector with length 3.}
}
\description{
This function returns Exact HWtest p and probability of all possible counts.
}
\examples{
xx<-ry_hweExact(c(813,182,5))
xx$p.value
sum(xx$prob) # should be 1
}
\keyword{Hardy-Weinberg}
\keyword{equilibrium}
\keyword{exact}
\keyword{test,}

