% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ry_momentX.R
\name{ry_momentX}
\alias{ry_momentX}
\title{Arbitrary moment}
\usage{
ry_momentX(x, order = 1, center = FALSE)
}
\arguments{
\item{x}{is a numeric vector.}

\item{order}{is order of moment.}

\item{center}{is logical; when TRUE, moment is around mean
and when FALSE, around zero}
}
\description{
This function returns order-th sample moment.
}
\examples{
n <- 10
x <- runif(n)
out2 <- ry_momentX(x,order=2,center=TRUE)
}
\keyword{complex}
\keyword{conformal}
\keyword{function,}
\keyword{transformation}

