% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ry_powerSetHasse.R
\name{ry_powerSetHasse}
\alias{ry_powerSetHasse}
\title{Draw a Hasse Diagram}
\usage{
ry_powerSetHasse(k = 4)
}
\arguments{
\item{k}{is an integer..}
}
\description{
This function draws a Hasse diagram..
}
\examples{
ry_powerSetHasse(4)
}
\keyword{Hasse}
\keyword{powerSet}

