% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ry_sample_fx.R
\name{ry_sample_fx}
\alias{ry_sample_fx}
\title{A Sample Function}
\usage{
ry_sample_fx(n)
}
\arguments{
\item{n}{is integer for number of randam values.}
}
\description{
This is a toy function
}
\examples{
n <- 100
ry_sample_fx(n)
}
\keyword{toy}

