% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ry_sample_fx2.R
\name{ry_sample_fx2}
\alias{ry_sample_fx2}
\title{A Sample Function2}
\usage{
ry_sample_fx2(n, m)
}
\arguments{
\item{n}{is integer for number of randam values.}
}
\description{
This is a toy function
}
\examples{
n <- 100
m <- 3
ry_sample_fx2(n,m)
}
\keyword{toy}

