% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_complex_color.R
\name{my_complex_color}
\alias{my_complex_color}
\title{Color with complex numbers}
\usage{
my_complex_color(z, int0 = 0.6, sat0 = 0.3, int1 = 1, sat1 = 1)
}
\arguments{
\item{z}{is a complex value vector.}

\item{sat0}{a parameter.}

\item{sat1}{a parameter.}

\item{init0}{a parameter.}

\item{init1}{a paramter.}
}
\description{
This function gives rgb color based on complex numbers.
}
\examples{
x <- seq(from=0,to=2*pi,len=1000)
xx <- expand.grid(x,x)
z <- xx[,1]+1i * xx[,2]
log.z <- log(z)
plot(z,col=my_complex_color(log.z,0.1,0.1,1,1),pch=20)
}
\keyword{complex}
\keyword{number}

